Se agrega ejemplo de form

+++ ChallengeDiWork/Modelo/Desperfecto.cs
@@ -5,15 +5,15 @@ using System.Text;
 using System.Threading.Tasks;
 
 namespace ChallengeDiWork.Modelo
+
 {
     public class Desperfecto
     {
         public string Descripcion { get; set; }
         public decimal ManoDeObra { get; set; }
-        public int Tiempo { get; set; }
+        public decimal Tiempo { get; set; }
         public Vehiculo Vehiculo { get; set; }
         public List<Repuesto> Repuestos { get; set; }
-
 
     }
 }
